resolve PR comments and remove redundant files
