--- final_project/datautils/stats.py
@@ -1,31 +1,0 @@
-import statistics as st
-import networkx as nx
-
-
-def calc_stats(graph_degrees: nx.Graph):
-    """
-    Calculate mean, median, mode, min and max of given graph degrees
-
-    Returns
-    -------
-    Namespace
-        Namespace with statistics
-    """
-
-    class Namespace:
-        def __init__(self, **kwargs):
-            self.__dict__.update(**kwargs)
-
-        def __str__(self):
-            return str(self.__dict__)
-
-    degrees = [
-        elem[1] for elem in sorted(graph_degrees, key=lambda x: x[1], reverse=True)
-    ]
-    return Namespace(
-        mean=st.mean(degrees),
-        median=st.median(degrees),
-        mode=st.mode(degrees),
-        min=min(degrees),
-        max=max(degrees),
-    )
