--- final_project/utils.py
@@ -1,3 +1,0 @@
-def print_header(title: str):
-    """Prints pretty header"""
-    print("="*20 + f" {title} " + "="*20 + "\n")
